Small warnings fix

+++ Bidou-core/src/graphics/buffers/vertexarray.cpp
@@ -7,7 +7,7 @@ namespace bidou { namespace graphics {
 	}
 
 	VertexArray::~VertexArray() {
-		for (int i = 0; i < m_Buffers.size(); i++)
+		for (size_t i = 0; i < m_Buffers.size(); i++)
 			delete m_Buffers[i];
 	}
